+++ pktwallet/prompt/prompt.go
@@ -15,7 +15,7 @@ import (
 	"github.com/pkt-cash/pktd/btcutil/er"
 
 	"github.com/pkt-cash/pktd/btcutil/hdkeychain"
-	"github.com/pkt-cash/pktd/pktwallet/internal/legacy/keystore"
+	"github.com/pkt-cash/pktd/pktwallet/legacy/keystore"
 	"github.com/pkt-cash/pktd/pktwallet/wallet/seedwords"
 	"golang.org/x/crypto/ssh/terminal"
 )
