+++ pktwallet/waddrmgr/scoped_manager.go
@@ -11,7 +11,7 @@ import (
 	"github.com/pkt-cash/pktd/btcutil"
 	"github.com/pkt-cash/pktd/btcutil/hdkeychain"
 	"github.com/pkt-cash/pktd/chaincfg"
-	"github.com/pkt-cash/pktd/pktwallet/internal/zero"
+	"github.com/pkt-cash/pktd/pktwallet/zero"
 	"github.com/pkt-cash/pktd/pktwallet/walletdb"
 )
