+++ pktwallet/walletsetup.go
@@ -21,9 +21,9 @@ import (
 	"github.com/pkt-cash/pktd/btcec"
 	"github.com/pkt-cash/pktd/btcutil"
 	"github.com/pkt-cash/pktd/chaincfg"
-	"github.com/pkt-cash/pktd/pktwallet/internal/legacy/keystore"
-	"github.com/pkt-cash/pktd/pktwallet/internal/prompt"
-	"github.com/pkt-cash/pktd/pktwallet/internal/zero"
+	"github.com/pkt-cash/pktd/pktwallet/legacy/keystore"
+	"github.com/pkt-cash/pktd/pktwallet/prompt"
+	"github.com/pkt-cash/pktd/pktwallet/zero"
 	"github.com/pkt-cash/pktd/pktwallet/waddrmgr"
 	"github.com/pkt-cash/pktd/pktwallet/wallet"
 	"github.com/pkt-cash/pktd/pktwallet/wallet/seedwords"
