+++ pktwallet/zero/array.go
