+++ pktwallet/zero/slice.go
