+++ pktwallet/zero/zero_test.go
@@ -10,7 +10,7 @@ import (
 	"testing"
 
 	"github.com/pkt-cash/pktd/btcutil/er"
-	. "github.com/pkt-cash/pktd/pktwallet/internal/zero"
+	. "github.com/pkt-cash/pktd/pktwallet/zero"
 )
 
 func makeOneBytes(n int) []byte {
